requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms>=2.3
crispy-bootstrap5==2025.6
whitenoise==6.6.0
python-dotenv==1.0.0
